Move all price querying logic from history.py to cryptocompare.py
